% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_melody_features.R
\name{install_melodyfeatures}
\alias{install_melodyfeatures}
\title{Install Python dependencies}
\usage{
install_melodyfeatures(envname = "r-melodyfeatures", force = FALSE)
}
\description{
Creates the dedicated melodyfeatures Python environment and
installs required packages.
}
